cleaning & prep for auto-sync by known user feature
