+++ submodules/SlackHelpers/handleButtonClickDelete.js
@@ -1,0 +1,7 @@
+const handleButtonClickDelete = (bot, trigger) => {
+  bot.replyInteractive(trigger, {
+    delete_original: true
+  });
+};
+
+module.exports = handleButtonClickDelete;
