+++ submodules/SlackHelpers/handleButtonClickHide.js
@@ -1,0 +1,7 @@
+const handleButtonClickHide = (bot, trigger) => {
+  const reply = trigger.original_message;
+  reply.attachments = [];
+  bot.replyInteractive(trigger, reply);
+};
+
+module.exports = handleButtonClickHide;
